# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/events/sessions/start": {
      "post": {
        "operationId": "EventsController_startSession",
        "summary": "Iniciar nova sessão de eventos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sessão criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/sessions/{sessionId}/events": {
      "post": {
        "operationId": "EventsController_addEvent",
        "summary": "Adicionar evento à sessão",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Evento adicionado com sucesso"
          },
          "400": {
            "description": "Sessão não está ativa"
          },
          "404": {
            "description": "Sessão não encontrada"
          }
        },
        "tags": [
          "Events"
        ]
      },
      "get": {
        "operationId": "EventsController_getSessionEvents",
        "summary": "Listar eventos da sessão",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Sessão não encontrada"
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/sessions/{sessionId}/commit": {
      "post": {
        "operationId": "EventsController_commitSession",
        "summary": "Finalizar e processar sessão",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sessão processada com sucesso"
          },
          "400": {
            "description": "Sessão não está ativa"
          },
          "404": {
            "description": "Sessão não encontrada"
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/sessions/{sessionId}/rollback": {
      "post": {
        "operationId": "EventsController_rollbackSession",
        "summary": "Cancelar sessão",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sessão cancelada com sucesso"
          },
          "404": {
            "description": "Sessão não encontrada"
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/sessions/{sessionId}/status": {
      "get": {
        "operationId": "EventsController_getSessionStatus",
        "summary": "Obter status da sessão",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status da sessão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "Sessão não encontrada"
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/metrics": {
      "get": {
        "operationId": "MetricsController_getMetrics",
        "summary": "Obter métricas do sistema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Métricas do sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Metrics"
        ]
      }
    }
  },
  "info": {
    "title": "Events Module API",
    "description": "API para gerenciamento de eventos com sessões transacionais",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Events",
      "description": ""
    },
    {
      "name": "Metrics",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "StartSessionDto": {
        "type": "object",
        "properties": {
          "filialId": {
            "type": "string",
            "description": "ID da filial/matriz"
          },
          "filialCNPJ": {
            "type": "string",
            "description": "CNPJ da filial/matriz"
          },
          "ambiente": {
            "type": "string",
            "description": "Ambiente (ex: production, staging, development)"
          },
          "sender": {
            "type": "string",
            "description": "Identificador do remetente/sistema"
          }
        },
        "required": [
          "filialId",
          "filialCNPJ",
          "ambiente",
          "sender"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "ID da sessão"
          },
          "message": {
            "type": "string",
            "description": "Mensagem de status"
          }
        },
        "required": [
          "sessionId",
          "message"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Dados do evento"
          },
          "method": {
            "type": "string",
            "description": "Método/ação executada"
          },
          "className": {
            "type": "string",
            "description": "Nome da classe/entidade"
          },
          "unico": {
            "type": "string",
            "description": "Identificador único do registro"
          },
          "filialId": {
            "type": "number",
            "description": "ID da filial"
          },
          "filialCnpj": {
            "type": "string",
            "description": "CNPJ da filial"
          }
        },
        "required": [
          "data",
          "method",
          "className",
          "unico",
          "filialId",
          "filialCnpj"
        ]
      },
      "SessionStatusDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "ID da sessão"
          },
          "status": {
            "type": "string",
            "description": "Status da sessão",
            "enum": [
              "ACTIVE",
              "PROCESSING",
              "COMMITTED",
              "FAILED",
              "ROLLED_BACK"
            ]
          },
          "eventCount": {
            "type": "number",
            "description": "Quantidade de eventos"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação"
          }
        },
        "required": [
          "sessionId",
          "status",
          "eventCount",
          "createdAt"
        ]
      },
      "Event": {
        "type": "object",
        "properties": {}
      },
      "MetricsResponseDto": {
        "type": "object",
        "properties": {
          "totalEvents": {
            "type": "number",
            "description": "Total de eventos processados"
          },
          "successRate": {
            "type": "number",
            "description": "Taxa de sucesso (%)"
          },
          "averageDuration": {
            "type": "number",
            "description": "Duração média de processamento (ms)"
          },
          "eventsPerMinute": {
            "type": "number",
            "description": "Eventos por minuto"
          },
          "deadLetterCount": {
            "type": "number",
            "description": "Quantidade de mensagens na DLQ"
          },
          "recentFailures": {
            "description": "Falhas recentes",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "totalEvents",
          "successRate",
          "averageDuration",
          "eventsPerMinute",
          "deadLetterCount",
          "recentFailures"
        ]
      }
    }
  }
}